// types/media.ts
export interface InstagramPost {
  id: string;
  imageUrl: string;
  caption: string;
}

export interface YouTubeVideo {
  id: string;
  videoId: string;
  title: string;
  description: string;
  thumbnail: string;
}

export interface Video {
  id: string;
  url: string;
  thumbnail: string;
  title: string;
  description: string;
}